requirements/local.txt
amqp==2.5.0
Babel==2.7.0
billiard==3.6.0.0
celery==4.3.0
certifi==2019.3.9
chardet==3.0.4
Django==2.2.1
django-crispy-forms==1.7.2
django-debug-toolbar==1.11
django-discover-runner==1.0
django-environ==0.4.5
django-extensions==2.1.7
django-recaptcha2==1.4.1
django-role-permissions==2.2.1
idna==2.8
kombu==4.6.0
money==1.3.0
Pillow==6.0.0
pkg-resources==0.0.0
psycopg2-binary==2.8.2
pytz==2019.1
requests==2.22.0
six==1.12.0
sqlparse==0.3.0
urllib3==1.25.3
vine==1.3.0
whitenoise==4.1.2
